bot/requirements.txt
requests
beautifulsoup4
sqlalchemy
psycopg2-binary
python-dotenv